Mapseeding work, extract interface of the enumerator and added game builder
Added a GameBuilder class to build the game and the map before handing it over to the actual game.
Extracted the interface from SimpleAreaEnumertor to make room for a random one later.
Added an additional test and rewrote mapseeding in structure and interfaces (as in what comes in and goes out)